Reference

VCF.Header

VCF.Header

VCF.MetaInfo

VCF.MetaInfo

VCF.MetaInfo

VCF.Reader

VCF.Record

VCF.Record

VCF.Record

VCF.Writer

BioCore.header

VCF.alt

VCF.chrom

VCF.filter

VCF.format

VCF.genotype

VCF.genotype

VCF.genotype

VCF.id

VCF.info

VCF.info

VCF.infokeys

VCF.pos

VCF.qual

VCF.ref

VCF.Header — Method

VCF.Header(metainfo::Vector, sampleID::Vector)

Create a VCF header with metainfo and sampleID.

VCF.Header — Method

VCF.Header()

Create an empty VCF header.

VCF.MetaInfo — Method

VCF.MetaInfo(str::AbstractString)

Create a VCF metainfo from str containing a VCF header line. This function verifies the format and indexes fields for accessors.

VCF.MetaInfo — Method

VCF.MetaInfo(data::Vector{UInt8})

Create a VCF metainfo from data containing a VCF header line. This function verifies the format and indexes fields for accessors. Note that the ownership of data is transferred to a new metainfo object.

VCF.MetaInfo — Method

VCF.MetaInfo()

Create an unfilled VCF metainfo.

VCF.Reader — Method

VCF.Reader(input::IO)

Create a data reader of the VCF file format.

Arguments

input: data source

VCF.Record — Method

VCF.Record(str::AbstractString)

Create a VCF object from str containing a VCF record. This function verifies the format and indexes fields for accessors.

VCF.Record — Method

VCF.Record(data::Vector{UInt8})

Create a VCF object from data containing a VCF record. This function verifies the format and indexes fields for accessors. Note that the ownership of data is transferred to a new record.

VCF.Record — Method

VCF.Record()

Create an unfilled VCF record.

VCF.Writer — Method

VCF.Writer(output::IO, header::VCF.Header)

Create a data writer of the VCF file format.

Arguments

output: data sink

header: VCF header object

BioCore.header — Method

header(reader::VCF.Reader)::VCF.Header

Get the header of reader.

VCF.alt — Method

alt(record::Record)::Vector{String}

Get the alternate bases of record.

VCF.chrom — Method

chrom(record::Record)::String

Get the chromosome name of record.

VCF.filter — Method

filter(record::Record)::Vector{String}

Get the filter status of record.

VCF.format — Method

format(record::Record)::Vector{String}

Get the genotype format of record.

VCF.genotype — Method

genotype(record::Record, indexes, keys)

Get the genotypes in record that match indexes and keys. indexes and keys can be either a scalar or a vector value. Trailing fields that are dropped are filled with ".".

VCF.genotype — Method

genotype(record::Record, index::Integer)::Vector{String}

Get the genotypes of the index-th individual in record. This is effectively equivalent to genotype(record)[index] but more efficient.

VCF.genotype — Method

genotype(record::Record)::Vector{Vector{String}}

Get the genotypes of record.

VCF.id — Method

id(record::Record)::Vector{String}

Get the identifiers of record.

VCF.info — Method

info(record::Record, key::String)::String

Get the additional information of record with key. Keys without corresponding values return an empty string.

VCF.info — Method

info(record::Record)::Vector{Pair{String,String}}

Get the additional information of record.

VCF.infokeys — Method

infokeys(record::Record)::Vector{String}

Get the keys of the additional information of record. This function returns an empty vector when the INFO field is missing.

VCF.pos — Method

pos(record::Record)::Int

Get the reference position of record.

VCF.qual — Method

qual(record::Record)::Float64

Get the quality score of record.

VCF.ref — Method

ref(record::Record)::String

Get the reference bases of record.