Reference

VariantCallFormat.Header

VariantCallFormat.Header

VariantCallFormat.MetaInfo

VariantCallFormat.MetaInfo

VariantCallFormat.MetaInfo

VariantCallFormat.Reader

VariantCallFormat.Record

VariantCallFormat.Record

VariantCallFormat.Record

VariantCallFormat.Writer

BioCore.header

VariantCallFormat.alt

VariantCallFormat.chrom

VariantCallFormat.filter

VariantCallFormat.format

VariantCallFormat.genotype

VariantCallFormat.genotype

VariantCallFormat.genotype

VariantCallFormat.id

VariantCallFormat.info

VariantCallFormat.info

VariantCallFormat.infokeys

VariantCallFormat.pos

VariantCallFormat.qual

VariantCallFormat.ref

VariantCallFormat.Header — Method

VCF.Header(metainfo::Vector, sampleID::Vector)

Create a VCF header with metainfo and sampleID.

VariantCallFormat.Header — Method

VCF.Header()

Create an empty VCF header.

VariantCallFormat.MetaInfo — Method

VCF.MetaInfo(str::AbstractString)

Create a VCF metainfo from str containing a VCF header line. This function verifies the format and indexes fields for accessors.

VariantCallFormat.MetaInfo — Method

VCF.MetaInfo(data::Vector{UInt8})

Create a VCF metainfo from data containing a VCF header line. This function verifies the format and indexes fields for accessors. Note that the ownership of data is transferred to a new metainfo object.

VariantCallFormat.MetaInfo — Method

VCF.MetaInfo()

Create an unfilled VCF metainfo.

VariantCallFormat.Reader — Method

VCF.Reader(input::IO)

Create a data reader of the VCF file format.

Arguments

input: data source

VariantCallFormat.Record — Method

VCF.Record(str::AbstractString)

Create a VCF object from str containing a VCF record. This function verifies the format and indexes fields for accessors.

VariantCallFormat.Record — Method

VCF.Record(data::Vector{UInt8})

Create a VCF object from data containing a VCF record. This function verifies the format and indexes fields for accessors. Note that the ownership of data is transferred to a new record.

VariantCallFormat.Record — Method

VCF.Record()

Create an unfilled VCF record.

VariantCallFormat.Writer — Method

VCF.Writer(output::IO, header::VCF.Header)

Create a data writer of the VCF file format.

Arguments

output: data sink

header: VCF header object

BioCore.header — Method

header(reader::VCF.Reader)::VCF.Header

Get the header of reader.

VariantCallFormat.alt — Method

alt(record::Record)::Vector{String}

Get the alternate bases of record.

VariantCallFormat.chrom — Method

chrom(record::Record)::String

Get the chromosome name of record.

VariantCallFormat.filter — Method

filter(record::Record)::Vector{String}

Get the filter status of record.

VariantCallFormat.format — Method

format(record::Record)::Vector{String}

Get the genotype format of record.

VariantCallFormat.genotype — Method

genotype(record::Record, indexes, keys)

Get the genotypes in record that match indexes and keys. indexes and keys can be either a scalar or a vector value. Trailing fields that are dropped are filled with ".".

VariantCallFormat.genotype — Method

genotype(record::Record, index::Integer)::Vector{String}

Get the genotypes of the index-th individual in record. This is effectively equivalent to genotype(record)[index] but more efficient.

VariantCallFormat.genotype — Method

genotype(record::Record)::Vector{Vector{String}}

Get the genotypes of record.

VariantCallFormat.id — Method

id(record::Record)::Vector{String}

Get the identifiers of record.

VariantCallFormat.info — Method

info(record::Record, key::String)::String

Get the additional information of record with key. Keys without corresponding values return an empty string.

VariantCallFormat.info — Method

info(record::Record)::Vector{Pair{String,String}}

Get the additional information of record.

VariantCallFormat.infokeys — Method

infokeys(record::Record)::Vector{String}

Get the keys of the additional information of record. This function returns an empty vector when the INFO field is missing.

VariantCallFormat.pos — Method

pos(record::Record)::Int

Get the reference position of record.

VariantCallFormat.qual — Method

qual(record::Record)::Float64

Get the quality score of record.

VariantCallFormat.ref — Method

ref(record::Record)::String

Get the reference bases of record.